Add 3-encoder forward method
Added a method to move the robot forward using input from 3 encoders (on the front two wheels and the back left wheel).
This functionality was added to the forward_enc() method, taking the same tiles argument as the previous forward() method. The forward() method now delegates to this new method, but the time-base movement method still exists with the name forward_time().
This commit is a proof of concept before other motions are edited to be controlled by encoder readings.