Add Vector Movement
Added vector movement, through the move() method. It takes three arguments, but is overloaded so the third is optional:
y_tiles: Number of tiles to move forward. Negative values move backward.
x_tiles: Number of tiles to move right. Negative values move left.
name: The name of the movement to display on the Driver Station. Defaults to "Move".