Add Camera Position Corrections
Add `setCameraPosition()` to set the position of the camera. This is stored in the private variables `cam_x`, `cam_y`, and `cam_z`. All measurements are now in mm and from the "reference" on the robot (the camera by default). Also, add an additional patch to fix the NullPointerException when recognizing objects with a webcam.